pyproject.toml
[project]
name = "ksp-anchor-keeper"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pyyaml>=6.0.2",
    "sfsutils>=1.1.1",
    "watchdog>=6.0.0",
]
